EquiTrack/requirements/base.txt
Django==1.7.10
bpython==0.12
django-braces==1.2.2
django-model-utils==2.3.1
logutils==0.3.3
psycopg2==2.5.1
dj-database-url==0.2.2
easy-thumbnails==1.4
celery==3.1.12
dealer==2.0.4
django-celery==3.1.16
django-celery-email==1.1.1
django-filer==0.9.12
#django-storages-redux==1.3
django-suit==0.2.14
django-suit-ckeditor==0.0.2
django-reversion==1.8.2
django-import-export==0.1.5
django-autocomplete-light==2.0.0a6
django-generic-links==0.2.0
jsonfield==1.0.3
django-logentry-admin==0.1.5
django-messages-extends==0.3.1
djrill==1.1.1
git+https://github.com/digi604/django-smart-selects.git
djangorestframework==3.2.3
djangorestframework-csv==1.3.3
djangorestframework-jwt==1.7.2
django-redis-cache==0.10.2
hiredis==0.1.2
boto==2.38.0
GDAL==1.10.0
Fiona==1.5.0
Shapely==1.5.7
cartodb==0.8.1
gunicorn==18.0
django-datetime-widget==0.9.3
django-leaflet==0.14.1
django-geojson==2.6.0
django-post_office==1.1.2
django-cors-headers==1.0.0
django-mathfilters==0.3.0
django-tenant-schemas==1.5.7
django-fsm==2.3.0
djangosaml2==0.13.0
tenant-schemas-celery==0.1.5
azure-storage==0.20.2

# testing libs
coverage==4.0
factory_boy==2.5.2
django-debug-toolbar==1.3.2

git+https://github.com/jschneier/django-storages.git#egg=django-storages
git+https://github.com/jamescw/django-paintstore.git
git+https://github.com/UNICEFLebanonInnovation/ActvityInfoPython.git
git+https://github.com/UNICEFLebanonInnovation/django-registration.git@0.9.2b9

